Extract FakeStdError into its own file since it's used in multiple places now
