+++ spec/inertia/partial_reload_spec.rb
@@ -1,19 +1,3 @@
-class FakeStdErr
-  attr_accessor :messages
-
-  def initialize
-    @messages = []
-  end
-
-  def write(msg)
-    @messages << msg
-  end
-
-  # Rails 5.0 + Ruby 2.6 require puts to be a public method
-  def puts(thing)
-  end
-end
-
 RSpec.describe 'partial reloads', type: :request do
   describe 'optimization guard rails' do
     context 'when a non-requested prop is defined as a value' do
