+++ spec/inertia/rspec_helper_spec.rb
@@ -1,21 +1,5 @@
 require_relative '../../lib/inertia_rails/rspec'
 
-class FakeStdErr
-  attr_accessor :messages
-
-  def initialize
-    @messages = []
-  end
-
-  def write(msg)
-    @messages << msg
-  end
-
-  # Rails 5.0 + Ruby 2.6 require puts to be a public method
-  def puts(thing)
-  end
-end
-
 RSpec.describe InertiaRails::RSpec, type: :request do
   describe 'correctly set up inertia tests with inertia: true', inertia: true do
     context 'with props' do
